fix: adminservice service add finishing
