+++ src/components/adminservice/edit/adminserviceEditSubservice.tsx
@@ -3,23 +3,22 @@ import IconDrag from "@/components/ui/IconDragAddAdmin";
 // ประกาศ interface สำหรับ props ของ AddSubService
 interface AddSubServiceProps {
   index: number;
-  subservice: { description: string; unit: string; pricePerUnit: number };
+  subservice: { description: string; unit: string; unit_price: number };
   deleteSubservice: (index: number) => void;
   updateSubservice: (
     index: number,
-    field: "description" | "unit" | "pricePerUnit",
+    field: "description" | "unit" | "unit_price",
     value: string | number
   ) => void;
   subServiceEmpty: boolean;
 }
 
-export function AddSubService({
+export function AdminserviceEditSubService({
   index,
   subservice,
   deleteSubservice,
   updateSubservice,
   subServiceEmpty,
-
 }: AddSubServiceProps) {
   return (
     <>
@@ -41,49 +40,39 @@ export function AddSubService({
               </span>
             )}
           </label>
-          {/* {subserviceEmpty && <p className="text-red-500">กรอกข้อมูลไม่ครบ</p>} */}
           <input
             type="text"
             id={`subserviceName-${index}`}
             value={subservice.description}
-            onChange={(e) => {
-              updateSubservice(index, "description", e.target.value);
-              // cancleActionEmpty("description", e.target.value)
-            }}
-            className="border border-gray-300 h-11 rounded-lg w-[422px] pl-5 "
-            />
-            {/* {index === 0 &&subserviceEmpty && (
-              <p className="text-red-500 text-sm  absolute bottom-0 font-medium">กรุณากรอกข้อมูลให้ถูกต้อง</p> */}
+            onChange={(e) =>
+              updateSubservice(index, "description", e.target.value)
+            }
+            className="border border-gray-300 h-11 rounded-lg w-[422px] pl-5"
+          />
           {index === 0 && subServiceEmpty && (
             <p className="text-red-500 text-sm  absolute bottom-0 font-medium">
               กรุณากรอกข้อมูลให้ถูกต้อง
             </p>
           )}
         </div>
-        <div className="flex flex-col py-6 relative">
+        <div className="flex flex-col py-6">
           <label
             htmlFor={`subservicePrice-${index}`}
-            className="text-sm text-gray-600 "
+            className="text-sm text-gray-600"
           >
             ค่าบริการ / 1 หน่วย
           </label>
-          {index === 0 && (
-            <span className="text-red-600 text-base absolute left-[115] top-5 font-medium">
-              *
-            </span>
-          )}
           <input
             type="number"
             id={`subservicePrice-${index}`}
-            // value={subservice.pricePerUnit}
+            value={subservice.unit_price}
             onChange={(e) => {
-              updateSubservice(index, "pricePerUnit", e.target.value);
-              // cancleActionEmpty("pricePerUnit", e.target.value)
+              updateSubservice(index, "unit_price", e.target.value);
             }}
             className="border border-gray-300 h-11 rounded-lg w-[240px] pl-5"
           />
         </div>
-        <div className="flex flex-col py-6 relative">
+        <div className="flex flex-col py-6">
           <label
             htmlFor={`subserviceUnit-${index}`}
             className="text-sm text-gray-600"
@@ -91,16 +80,15 @@ export function AddSubService({
             หน่วยการบริการ
           </label>
           {index === 0 && (
-            <span className="text-red-600 text-base absolute left-[94] top-5 font-medium">*</span>
+            <span className="text-red-600 text-base absolute left-[94] top-5 font-medium">
+              *
+            </span>
           )}
           <input
             type="text"
             id={`subserviceUnit-${index}`}
             value={subservice.unit}
-            onChange={(e) => {
-              updateSubservice(index, "unit", e.target.value);
-              // cancleActionEmpty("unit", e.target.value)
-            }}
+            onChange={(e) => updateSubservice(index, "unit", e.target.value)}
             className="border border-gray-300 h-11 rounded-lg w-[240px] pl-5"
           />
         </div>
@@ -114,3 +102,47 @@ export function AddSubService({
     </>
   );
 }
+
+
+// import IconDrag from "@/components/ui/IconDragAddAdmin";
+
+// export function AdminserviceEditSubService({
+//   index,
+//   subservice,
+//   deleteSubservice,
+//   updateSubservice,
+// }: {
+//   index: number;
+//   subservice: {
+//     description: string;
+//     unit: string;
+//     unit_price: number;
+//     subServiceEmpty: boolean;
+//   };
+
+//   deleteSubservice: (index: number) => void;
+
+//   updateSubservice: (
+//     index: number,
+//     field: "description" | "unit" | "unit_price",
+//     value: string | number
+//   ) => void;
+// }) {
+
+{/* <div className="flex flex-col py-6">
+<label
+  htmlFor={`subservicePrice-${index}`}
+  className="text-sm text-gray-600"
+>
+  ค่าบริการ / 1 หน่วย
+</label>
+<input
+  type="number"
+  id={`subservicePrice-${index}`}
+  value={subservice.unit_price !== null ? subservice.unit_price : ""}
+  onChange={(e) =>
+    updateSubservice(index, "unit_price", e.target.value)
+  }
+  className="border border-gray-300 h-11 rounded-lg w-[240px] pl-5"
+/>
+</div> */}
