+++ src/pages/admincategory/index.tsx
@@ -1,4 +1,4 @@
-import Adminsidebar from "@/components/admin/admin-sidebar";
+import { AdminSidebar } from "@/components/admin/admin-sidebar";
 import { supabase } from "@/utils/supabase";
 import { useEffect, useState } from "react";
 import { useServices } from "@/components/ServicesContext";
@@ -20,7 +20,7 @@ export default function AdminNavbar() {
     <>
       <div className="flex flex-row w-full">
         <div>
-          <Adminsidebar />
+          <AdminSidebar />
         </div>
         <div className="w-full flex flex-col">
           {/* navbar for admin page */}
