--- src/pages/adminservice/detail.tsx
@@ -1,5 +1,0 @@
-export default function testA() {
-    return(
-        <div>ABC</div>
-    )
-}
