+++ src/pages/adminservice/detail/[id].tsx
@@ -1,23 +1,8 @@
 import { AdminSidebar } from "@/components/admin/admin-sidebar";
 import { useEffect, useState } from "react";
-import { useServices } from "@/components/ServicesContext";
 import { useRouter } from "next/router";
-import { FormEvent } from "react";
 import axios from "axios";
-import IconPicture from "@/components/ui/IconPicture";
-import IconDrag from "@/components/ui/IconDragAddAdmin";
-import IconPlusDefaultColor from "@/components/ui/IconPluseDefaultColor";
-import { stringify } from "querystring";
-import {
-  Select,
-  SelectContent,
-  SelectItem,
-  SelectTrigger,
-  SelectValue,
-} from "@/components/ui/select";
-import IconWarning from "@/components/ui/Iconwarning";
-import IconX from "@/components/ui/IconX";
-import Image from "next/image";
+import { AdminserviceDetailNavbar } from "@/components/adminservice/detail/adminserviceDetailNavbar";
 
 interface SubService {
   description: string;
@@ -25,137 +10,44 @@ interface SubService {
   unit_price: number;
 }
 
-export default function AdminNavbar() {
+export default function AdminserviceDetailIndex() {
   const [inputSubservice, setInputSubservice] = useState<SubService[]>([]);
   console.log(
     "check subservice id and value of subservice when refresh window",
     inputSubservice
   );
-  const modifyInputSubservice = () => {
-    const updatedSubservices = inputSubservice.map((subservice) => ({
-      ...subservice,
-      unit_price: subservice.unit_price ?? 0, // กำหนดค่า unit_price เป็น 0 ถ้าเป็น null
-    }));
-    setInputSubservice(updatedSubservices);
-  };
   const [inputTitle, setInputTitle] = useState<string>("");
   const [inputCat, setInputCat] = useState<number | undefined>();
   // console.log("this test for receive value inputCat", inputCat);
   const [inputImage, setInputImage] = useState<File>();
   // console.log("check image when update", inputImage);
-  const [nameTopic, setNameTopic] = useState<String>("loading");
+  const [nameTopic, setNameTopic] = useState<string>("loading");
   const [URLimage, setURLimage] = useState<String>();
   // const [changeImage,setChangeImage] = useState<Boolean>(false)
-  const [showPopUpSubmit, setShowPopUpSubmit] = useState<Boolean>(false);
   const [showPopUpDeleteImg, setShowPopUpDeleteImg] = useState<Boolean>(false);
 
   const router = useRouter();
   const { id } = router.query;
 
-  const handleSubmit = async (e: any) => {
-    e.preventDefault();
-
-    const formData = new FormData();
-    // formData.append("service_id", id: Number)
-
-    const serviceId = Array.isArray(id) ? id[0] : id; // เลือกค่าตัวแรกจากอาร์เรย์ ถ้า id เป็นอาร์เรย์
-
-    if (serviceId) {
-      formData.append("service_id", serviceId.toString()); // แปลงเป็น string ถ้าไม่ใช่
-    }
-
-    formData.append("title", inputTitle);
-    if (inputCat) {
-      formData.append("category_id", inputCat.toString());
-    }
-
-    if (inputImage) {
-      formData.append("image", inputImage);
-    }
-
-    formData.append("subservices", JSON.stringify(inputSubservice));
-
-    console.log("FormData contents:");
-    let formDataObject: { [key: string]: any } = {};
-    for (let [key, value] of formData.entries()) {
-      formDataObject[key] = value;
-    }
-
-    console.log(
-      "test data for sent request by form data.append",
-      formDataObject
-    );
-
-    // Commented out API call
-    // try {
-    //   await axios.put(`/api/admin/management/edit/${id}`, formData, {
-    //     headers: { "Content-Type": "multipart/form-data" },
-    //   });
-    //   router.push("/adminservice");
-    //   console.log("fromdata2", formData);
-    // } catch (error) {
-    //   console.log(error);
-    // }
-  };
-
-  // // ตัวอย่างการเรียกใช้ modifyInputSubservice
-  // useEffect(() => {
-  //   modifyInputSubservice(); // เรียกใช้ฟังก์ชัน
-  // }, [inputSubservice]); // เรียกใช้เมื่อ inputSubservice เปลี่ยนแปลง
-
   return (
     <>
-      <form
-        onSubmit={handleSubmit}
-        onKeyDown={(e) => {
-          const target = e.target as HTMLElement; // Cast ให้เป็น HTMLElement
-          if (e.key === "Enter" && target.tagName !== "TEXTAREA") {
-            e.preventDefault(); // ป้องกันการกด Enter ยกเว้นใน <textarea>
-          }
-        }}
-      >
+      <form>
         <div className="flex flex-row w-full">
           <div>
             <AdminSidebar />
           </div>
           <div className="w-full flex flex-col">
             {/* navbar for admin page */}
-            <div className="flex flex-row items-center justify-between bg-white sticky top-0 h-20 px-10 py-5 min-w-[1200px] border-b  border-gray-300 z-10">
-              <div
-                className="flex flex-row items-center cursor-pointer"
-                onClick={() => router.push("/adminservice")}
-              >
-                <ArrowBack />
-                <div className="flex flex-col">
-                  <div className="text-xs">บริการ</div>
-                  <div className="text-xl font-medium">{nameTopic}</div>
-                </div>
-              </div>
-              <div className="h-full flex flex-row items-center gap-6 relative">
-                {/* <button
-                  className="bg-white text-defaultColor text-base h-full px-7 flex items-center gap-3 rounded-lg w-32 text-center justify-center border border-defaultColor"
-                  onClick={() => router.push("/adminservice")}
-                  type="button"
-                >
-                  ยกเลิก
-                </button> */}
-                <button
-                  className="bg-defaultColor text-white text-base h-full px-7 flex items-center gap-3 rounded-lg w-32 text-center justify-center"
-                  onClick={() => router.push(`/adminservice/edit/${id}`)}
-                >
-                  แก้ไข
-                </button>
-              </div>
-            </div>
-            <AdminserviceIndex
+            <AdminserviceDetailNavbar 
+            nameTopic={nameTopic} 
+            id={id} />
+            <AdminserviceDetailService
               setInputSubservice={setInputSubservice}
               setInputTitle={setInputTitle}
               setInputCat={setInputCat}
-              SetInputimage={setInputImage}
-              setURLimage={setURLimage}
+ 
               setNameTopic={setNameTopic}
-              setShowPopUpDeleteImg={setShowPopUpDeleteImg}
-              showPopUpDeleteImg={showPopUpDeleteImg}
+
               inputTitle={inputTitle}
             />
           </div>
@@ -165,21 +57,16 @@ export default function AdminNavbar() {
   );
 }
 
-export const AdminserviceIndex = ({
+export const AdminserviceDetailService = ({
   setInputSubservice,
   setInputTitle,
   setInputCat,
-  SetInputimage,
-  setURLimage,
   setNameTopic,
-  setShowPopUpDeleteImg,
-  showPopUpDeleteImg,
   inputTitle,
 }: any) => {
   const router = useRouter();
   const { id } = router.query;
 
-  const [category_id, setCategory_id] = useState<string>("");
   const [subservices, setSubservices] = useState<any[]>([]);
 
   // setInputSubservice(subservices)
@@ -201,84 +88,10 @@ export const AdminserviceIndex = ({
   const [createAt, setCreateAt] = useState<string>(new Date().toISOString());
   const [updateAt, setUpdateAt] = useState<string>(new Date().toISOString());
 
-  const addSubService = () => {
-    setSubservices((prevSubservices) => [
-      ...prevSubservices,
-      { description: "", unit: "", unit_price: null },
-    ]);
-  };
 
-  const deleteSubservice = (index: number) => {
-    const updatedSubservices = subservices.filter((_, idx) => idx !== index);
-    setSubservices(updatedSubservices);
-  };
 
-  const updateSubservice = (
-    index: number,
-    field: string,
-    value: string | number
-  ) => {
-    setSubservices((prevSubservices) => {
-      const updatedSubservices = [...prevSubservices];
-      updatedSubservices[index] = {
-        ...updatedSubservices[index],
-        [field]: value,
-      };
-      return updatedSubservices;
-    });
-  };
-
-  const handleInputTitle = (event: React.ChangeEvent<HTMLInputElement>) => {
-    // setServiceNameData(event.target.value);
-    setInputTitle(event.target.value);
-  };
-
-  const handleCategorySelect = (
-    event: React.ChangeEvent<HTMLSelectElement>
-  ) => {
-    const value = Number(event.target.value);
-    console.log("value for check inputCat", value);
-    setInputCat(value);
-  };
 
   const [preview, setPreview] = useState<string | null>(null);
-
-  const handleInputImg = (event: React.ChangeEvent<HTMLInputElement>) => {
-    const file = event.target.files?.[0];
-    console.log(event, "event for image");
-    if (file) {
-      const previewURL = URL.createObjectURL(file);
-      console.log("previewURL", previewURL);
-      setPreview(previewURL);
-      SetInputimage(file);
-    }
-  };
-
-  const handleDeleteImg = async () => {
-    setPreview(null);
-    setShowPopUpDeleteImg(false);
-    // try {
-    //   const response = await axios.delete(`/api/delete-image`, {
-    //     params: { serviceId: id },
-    //   });
-
-    //   console.log("Image deleted successfully", response.data);
-    // } catch (error) {
-    //   if (axios.isAxiosError(error)) {
-    //     // AxiosError: จัดการข้อผิดพลาดที่เกิดจาก axios
-    //     if (error.response) {
-    //       console.error("Error response from server:", error.response.data);
-    //     } else if (error.request) {
-    //       console.error("No response received:", error.request);
-    //     } else {
-    //       console.error("Unexpected Axios error:", error.message);
-    //     }
-    //   } else {
-    //     // Unknown error: จัดการกรณีที่ไม่ใช่ข้อผิดพลาดของ axios
-    //     console.error("An unknown error occurred:", error);
-    //   }
-    // }
-  };
 
   const fetchService = async () => {
     try {
@@ -358,29 +171,6 @@ export const AdminserviceIndex = ({
             <div className="border-gray-300 h-11 rounded-lg w-[433px]  text-black font-normal items-center flex">
               {serviceCategoryData}
             </div>
-            {/* <Select
-              onValueChange={(value: string) => {
-                handleCategorySelect({
-                  target: { value },
-                } as React.ChangeEvent<HTMLSelectElement>);
-              }}
-            >
-              <SelectTrigger className="w-[433px] text-base font-normal h-[44px] text-black">
-                <SelectValue placeholder={serviceCategoryData} />
-              </SelectTrigger>
-              <SelectContent>
-                {fetchDataCategories.map(
-                  (fetchDataCategories: any, index: any) => (
-                    <SelectItem
-                      key={fetchDataCategories.id.toString()}
-                      value={fetchDataCategories.id.toString()}
-                    >
-                      {fetchDataCategories.category}
-                    </SelectItem>
-                  )
-                )}
-              </SelectContent>
-            </Select> */}
           </div>
 
           <div className="flex flex-col gap-12">
@@ -399,30 +189,7 @@ export const AdminserviceIndex = ({
                         className="w-full h-full object-cover rounded-lg"
                       />
                     ) : (
-                      <div>
-                        <div className="w-full h-full flex items-center justify-center flex-col gap-3 p-4">
-                          <IconPicture />
-                          <div className="flex flex-row gap-2">
-                            <label
-                              htmlFor="file-upload"
-                              className="cursor-pointer text-blue-500"
-                            >
-                              อัพโหลดภาพ
-                            </label>
-                            <p className="text-gray-600 text-center ">หรือ</p>
-                            <p className="text-gray-600 text-center ">
-                              ลากและวางที่นี้
-                            </p>
-                          </div>
-                          <div className="flex flex-row gap-2 text-sm -mt-2 text-gray-600">
-                            <p>PNG,</p>
-                            <p>JPG</p>
-                            <p>ขนาดไม่เกิน</p>
-                            <p className="text-gray-600 text-center ">5MB</p>
-                          </div>
-                        </div>
-                        <div></div>
-                      </div>
+                      <div></div>
                     )}
                   </div>
                 </div>
@@ -443,27 +210,13 @@ export const AdminserviceIndex = ({
                   รายการบริการย่อย
                 </h1>
                 {subservices.map((subservice, index) => (
-                  <AddSubService
+                  <AdminserviceDetailSubService
                     key={index}
                     index={index}
                     subservice={subservice}
-                    deleteSubservice={deleteSubservice}
-                    updateSubservice={updateSubservice}
                   />
                 ))}
               </div>
-              {/* <div className="">
-                <button
-                  type="button"
-                  className="bg-white text-defaultColor text-base h-10 flex items-center justify-center gap-3 rounded-lg border border-defaultColor px-7 mt-3"
-                  onClick={addSubService}
-                >
-                  เพิ่มรายการ
-                  <span>
-                    <IconPlusDefaultColor />
-                  </span>
-                </button>
-              </div> */}
             </div>
             {/* กล่องล่าง */}
             <div className="h-px w-full bg-gray-300 -mt-8"></div>
@@ -506,52 +259,13 @@ export const AdminserviceIndex = ({
           </div>
         </div>
       </div>
-      {/* Popup for delete image */}
-      {showPopUpDeleteImg && (
-        <div className="fixed inset-0 flex flex-col items-center justify-center bg-black bg-opacity-50">
-          <div className="bg-white w-[360px] h-auto flex flex-col items-center rounded-xl p-4 gap-3">
-            <div className="w-full">
-              <div
-                className="w-full flex justify-end "
-                onClick={() => setShowPopUpDeleteImg(false)}
-              >
-                <IconX />
-              </div>
-              <div className="flex justify-center">
-                <IconWarning />
-              </div>
-            </div>
-            <h1 className="font-medium text-xl ">ยืนยันการลบรายการ ?</h1>
-            <h1 className="text-center text-gray-500">
-              {/* คุณต้องการลบรายการ ‘{serviceName}’ <br /> */}
-              ใช่หรือไม่
-            </h1>
-            <div className="flex flex-row gap-3">
-              <button
-                className="bg-defaultColor text-white w-28 py-2 rounded-lg font-medium"
-                onClick={handleDeleteImg}
-              >
-                ลบรายการ
-              </button>
-              <button
-                className="bg-white text-defaultColor border-[1px] border-defaultColor w-28 py-2 rounded-lg font-medium"
-                onClick={() => setShowPopUpDeleteImg(false)}
-              >
-                ยกเลิก
-              </button>
-            </div>
-          </div>
-        </div>
-      )}
     </>
   );
 };
 
-export function AddSubService({
+export function AdminserviceDetailSubService({
   index,
   subservice,
-  deleteSubservice,
-  updateSubservice,
 }: {
   index: number;
   subservice: {
@@ -560,12 +274,6 @@ export function AddSubService({
     unit_price: number;
   };
 
-  deleteSubservice: (index: number) => void;
-  updateSubservice: (
-    index: number,
-    field: "description" | "unit" | "unit_price",
-    value: string | number
-  ) => void;
 }) {
   return (
     <>
@@ -578,15 +286,6 @@ export function AddSubService({
           >
             ชื่อบริการ
           </label>
-          {/* <input
-            type="text"
-            id={`subserviceName-${index}`}
-            value={subservice.description}
-            onChange={(e) =>
-              updateSubservice(index, "description", e.target.value)
-            }
-            className="border border-gray-300 h-11 rounded-lg w-[422px] pl-5"
-          /> */}
           <div className="border-gray-300 h-11 rounded-lg w-[422px] items-center flex">
             {subservice.description}
           </div>
@@ -598,15 +297,6 @@ export function AddSubService({
           >
             ค่าบริการ / 1 หน่วย
           </label>
-          {/* <input
-            type="number"
-            id={`subservicePrice-${index}`}
-            value={subservice.unit_price !== null ? subservice.unit_price : ""}
-            onChange={(e) =>
-              updateSubservice(index, "unit_price", e.target.value)
-            }
-            className="border border-gray-300 h-11 rounded-lg w-[240px] pl-5"
-          /> */}
           <div className="border-gray-300 h-11 rounded-lg w-[240px] items-center flex">
             {subservice.unit_price}
           </div>
@@ -618,38 +308,11 @@ export function AddSubService({
           >
             หน่วยการบริการ
           </label>
-          {/* <input
-            type="text"
-            id={`subserviceUnit-${index}`}
-            value={subservice.unit}
-            onChange={(e) => updateSubservice(index, "unit", e.target.value)}
-            className="border border-gray-300 h-11 rounded-lg w-[240px] pl-5"
-          /> */}
           <div className="border-gray-300 h-11 rounded-lg w-[240px] items-center flex">
             {subservice.unit}
           </div>
         </div>
       </div>
     </>
-  );
-}
-
-export function ArrowBack() {
-  return (
-    <svg
-      width="40"
-      height="40"
-      viewBox="0 0 40 40"
-      fill="none"
-      xmlns="http://www.w3.org/2000/svg"
-    >
-      <path
-        d="M25.0007 31.6667L13.334 20L25.0007 8.33337"
-        stroke="#646C80"
-        strokeWidth="3"
-        strokeLinecap="round"
-        strokeLinejoin="round"
-      />
-    </svg>
   );
 }
